# Implement consumption_rates for individuals