--- a/inst/include/plant/internals.h
+++ b/inst/include/plant/internals.h
@@ -16,30 +16,36 @@ namespace plant {
 
 class Internals {
 public:
-  Internals(size_t s_size=0, size_t a_size=0)
+  Internals(size_t s_size=0, size_t a_size=0, size_t r_size=0)
       :
       state_size(s_size),
       aux_size(a_size),
+      resource_size(r_size),
       states(s_size, 0.0),
       rates(s_size, NA_REAL) ,
-      auxs(a_size, 0.0)
+      auxs(a_size, 0.0),
+      consumption_rates(r_size, NA_REAL)
     {}
   size_t state_size;
   size_t aux_size;
+  size_t resource_size;
 
 
   // Perhaps make these private so the () overloads below have some use
   std::vector<double> states;
   std::vector<double> rates;
   std::vector<double> auxs;
+  std::vector<double> consumption_rates;  // not quite as pithy
 
   double state(int i) const { return states[i]; }
   double rate(int i) const { return rates[i]; }
   double aux(int i) const { return auxs[i]; }
+  double consumption_rate(int i) const { return consumption_rates[i]; }
 
   void set_state(int i, double v) { states[i] = v; }
   void set_rate(int i, double v) { rates[i] = v; }
   void set_aux(int i, double v) { auxs[i] = v; }
+  void set_consumption_rate(int i, double v) { consumption_rates[i] = v; }
 
   void resize(size_t new_size, size_t new_aux_size) {
     state_size = new_size;
@@ -47,6 +53,11 @@ public:
     states.resize(new_size, 0.0);
     rates.resize(new_size, NA_REAL);
     auxs.resize(new_aux_size, 0.0);
+  }
+
+  void resize_consumption_rates(size_t new_resource_size) {
+    resource_size = new_resource_size;
+    consumption_rates.resize(new_resource_size, NA_REAL);
   }
 };