--- a/inst/include/plant/models/ff16_environment.h
+++ b/inst/include/plant/models/ff16_environment.h
@@ -55,10 +55,36 @@ public:
     canopy.r_init_interpolators(state);
   }
 
-  virtual void compute_rates() {
+  virtual void compute_rates(std::vector<double> const& resource_depletion) {
+    double infiltration;
+    double net_flux;
+
+    double drainage_multiplier = 0.1; // experimental only;
+
+    std::cout << "Time: " << time << std::endl;
+
+    // treat each soil layer as a separate resource pool
     for (size_t i = 0; i < vars.state_size; i++) {
-      vars.set_rate(i, extrinsic_driver_evaluate("rainfall", time) / (i+1));
+
+      // initial representation of drainage; to be improved
+      if(i == 0) {
+        infiltration = extrinsic_drivers["rainfall"].eval(time);
+      } else {
+        infiltration = std::max(vars.state(i - 1), 0.0) * drainage_multiplier;
+      }
+
+      // ecologically, soil water shouldn't go below zero
+      // truncating at zero until such a model is implemented
+      double drainage_rate = std::max(vars.state(i), 0.0) * drainage_multiplier;
+
+      net_flux = infiltration - resource_depletion[i] - drainage_rate;
+      vars.set_rate(i, net_flux);
+
+      std::cout << std::setprecision(2) << std::fixed;
+      std::cout << "Soil layer: " << i << "; infil. rate: " << infiltration << "; extraction rate: " << resource_depletion[i] << "; net flux: " << net_flux << "; water balance " << vars.state(i) << std::endl;
     }
+
+    std::cout << "\n " << std::endl;
   }
 
   std::vector<double> get_soil_water_state() const {