--- a/inst/include/plant/models/ff16_environment.h
+++ b/inst/include/plant/models/ff16_environment.h
@@ -5,7 +5,6 @@
 #include <plant/environment.h>
 #include <plant/canopy.h>
 #include <plant/interpolator.h>
-#include <iostream>
 
 using namespace Rcpp;