--- a/tests/testthat/test-strategy-ff16w.R
+++ b/tests/testthat/test-strategy-ff16w.R
@@ -13,16 +13,18 @@ test_that("FF16w Environment", {
   # Initialised with no water, no inflow
   expect_equal(e$soil$states, 0)
   
-  e$compute_rates()
-  expect_equal(e$soil$rates, 1) # default rainfall is now y = 1
+  e$compute_rates(c(0.4))
+  # 1 - 0.4 - 0.0*0.1 = 0.6
+  expect_equal(e$soil$rates, 0.6) # default rainfall is now y = 1
   
   # Make it rain
   x <- seq(0, 9, 1)
   y <- rep(5, 10)
   e$set_extrinsic_driver("rainfall", x, y)
   expect_equal(e$soil$states, 0)
-  e$compute_rates()
-  expect_equal(e$soil$rates, 5)
+  e$compute_rates(c(0.4))
+  # 5 - 0.4 - 0.0*0.1
+  expect_equal(e$soil$rates, 4.6)
   
   # Water logged
   e$set_soil_water_state(100)
@@ -34,8 +36,11 @@ test_that("FF16w Environment", {
                         rainfall = 10)
   
   expect_equal(e$soil$states, 0)
-  e$compute_rates()
-  expect_equal(e$soil$rates, 10)
+  e$compute_rates(c(0.4))
+  # 10 - 0.4 - 0.0*0.1
+  # weird because setting the soil water state to 100 doesn't affect 
+  # the rate at all since its still soil_layer 0
+  expect_equal(e$soil$rates, 9.6)
 
   e <- make_environment("FF16w", soil_number_of_depths = 2)
   
@@ -76,10 +81,9 @@ test_that("Rainfall spline basic run", {
   
   expect_equal(out$patch$environment$ode_size, 10)
   
-  # model is currently 1/z * int(0, t)(f(x)), where z is depth, 
-  # x is time, f is spline, t is number of years
+  #infiltration0 <- out$time^2
+  #resource_depletion0 <- 2 * area_leaf? * density
   
-  # check the rates are correct, ie 105^2/depth
   expect_equal(out$patch$environment$soil$rates, 
                sapply(seq(1, 10), function(x) { (out$time^2)/x }),
                tolerance = 1e-7)