… R/RcppR6.R
… inst/RcppR6_classes.yml
… scripts/test-profiler.R