@devmitch 2022-01-17
I have added a script to benchmark the test script in the OP under `R/benchmark.R`. The implementation on this branch gave the following results:
```
> run_resource_consumption_benchmarks(10)
Running resource consumption benchmarks`
Running with:
  soil_layers
1          10
2         100
3        1000
# A tibble: 3 × 14
  expression soil_layers      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory               time            gc               
  <bch:expr>       <dbl> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list>               <list>          <list>           
1 scm                 10   46.6ms   46.7ms     21.3     17.1MB    21.3      5     5      235ms <NULL> <Rprofmem [670 × 3]> <bench_tm [10]> <tibble [10 × 3]>
2 scm                100   56.8ms   58.2ms     17.3     17.1MB    11.5      6     4      347ms <NULL> <Rprofmem [671 × 3]> <bench_tm [10]> <tibble [10 × 3]>
3 scm               1000  154.7ms  157.7ms      6.36    17.1MB     4.24     6     4      943ms <NULL> <Rprofmem [671 × 3]> <bench_tm [10]> <tibble [10 × 3]>
```

whilst the implementation on #328 yields:
```
> run_resource_consumption_benchmarks(10)
Running resource consumption benchmarks`
Running with:
  soil_layers
1          10
2         100
3        1000
# A tibble: 3 × 14
  expression soil_layers      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory               time            gc               
  <bch:expr>       <dbl> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list>               <list>          <list>           
1 scm                 10   86.8ms     87ms     11.5     17.1MB    11.5      5     5   436.27ms <NULL> <Rprofmem [670 × 3]> <bench_tm [10]> <tibble [10 × 3]>
2 scm                100  151.1ms    155ms      6.52    17.1MB     6.52     5     5   766.81ms <NULL> <Rprofmem [671 × 3]> <bench_tm [10]> <tibble [10 × 3]>
3 scm               1000  685.5ms    691ms      1.45    17.1MB     2.17     4     6      2.76s <NULL> <Rprofmem [671 × 3]> <bench_tm [10]> <tibble [10 × 3]>
```

... which is surprisingly slower for 1000 soil layers (and even 100 layers), although this number is probably unrealistic.