@aornugent 2022-01-17
Thanks Mitch! Can you please add `layers = 0` for reference? I went to compare against the [Github Actions benchmark for `develop`](https://github.com/traitecoevo/plant/runs/4515329520) but it turns out your machine is much quicker ;)