@devmitch 2022-01-18
> Thanks Mitch! Can you please add `layers = 0` for reference? I went to compare against the [Github Actions benchmark for `develop`](https://github.com/traitecoevo/plant/runs/4515329520) but it turns out your machine is much quicker ;)

https://github.com/traitecoevo/plant/blob/930b9c9679859727b8b2ac36995f360a9dc0838b/R/ff16w.R#L77