@dfalster 2022-01-18
How about 1 layer then. Close to 0