@devmitch 2022-01-26
While looking into a segfault bug in the `test-cohort.R` tests, I noticed that while `consumption_rates` were being correctly resized for patches with `resize_consumption_rates()`, this was not being done when creating a single individual, unlike the other vectors which were being resized correctly in the Individual constructor:
```cpp
Individual(strategy_type_ptr s) : strategy(s) {
    if (strategy->aux_index.size() != s->aux_size()) {
      strategy->refresh_indices();
    }
    vars.resize(strategy_type::state_size(), s->aux_size()); // = Internals(strategy_type::state_size());
    set_state("height", strategy->height_0);

    vars.resize_consumption_rates(vars.state_size); // NEW - now correctly sets the size of consumption_rates
  }
```
The tests now pass for `test-cohort.R`, but got me wondering - what is the point of the `resource_size` member variable in `Internals` if it is always going to be the same as `state_size`, another member variable in the same class? 

For example, in `Patch`, we resize the consumption rates using `environment.ode_size()`:
https://github.com/traitecoevo/plant/blob/abdea4cc4aafc09c87d1d988a381d0430b6102e9/inst/include/plant/patch.h#L133
which is simply `vars.state_size()`:
https://github.com/traitecoevo/plant/blob/abdea4cc4aafc09c87d1d988a381d0430b6102e9/inst/include/plant/environment.h#L30

Thus, `resize_consumption_rates` only ever sets `resource_size` to `state_size`:
https://github.com/traitecoevo/plant/blob/abdea4cc4aafc09c87d1d988a381d0430b6102e9/inst/include/plant/internals.h#L58-L60

The only time I can see them being different is when a `resize()` is done after `resize_consumption_rates` - maybe I'm getting confused, but I would like some clarity on this.