@aornugent 2022-01-26
Hey @devmitch - thanks for the investigation. There's a few things to unpack here.

1. Individuals and environments have *separate* `vars` objects. For FF16w: `individual.state_size() = 5` (+2 if it's a cohort) and `environment.vars.ode_size() = n_soil_layers`. 
2. `resize_consumption_rates` is handled separately because individuals do not have access to the environment at construction. So when creating a patch, the individuals are instantiated using a strategy object, and then consumption rates are resized from the patch constructor which does have access to information about the environment.
3. Ideally, we'd keep this separation and allow an individual to be created without an environment and without any consumption rate variables when working outside the Patch framework. This maintains the modularity of the code desribed in the tests. I tend to think of a Patch as the union of a strategy (embodied by individuals or cohorts) and an environment.

It's a shame that the tests fail without. This is a shot in the dark, but do they pass if the default consumption rates size is set to zero? Keep the questions coming, it's a bit curly.